chore: refactor app provider
