+++ packages/frontend/src/hooks/health/useHealth.tsx
@@ -1,0 +1,43 @@
+import { ApiError, ApiHealthResults, HealthState } from '@lightdash/common';
+import React, { useEffect } from 'react';
+import { useQuery } from 'react-query';
+import { lightdashApi } from '../../api';
+import { AppToaster } from '../../components/AppToaster';
+
+const getHealthState = async () =>
+    lightdashApi<ApiHealthResults>({
+        url: `/health`,
+        method: 'GET',
+        body: undefined,
+    });
+
+const useHealth = () => {
+    const health = useQuery<HealthState, ApiError>({
+        queryKey: 'health',
+        queryFn: getHealthState,
+    });
+
+    useEffect(() => {
+        if (health.error) {
+            const [first, ...rest] = health.error.error.message.split('\n');
+            AppToaster.show(
+                {
+                    intent: 'danger',
+                    message: (
+                        <div>
+                            <b>{first}</b>
+                            <p>{rest.join('\n')}</p>
+                        </div>
+                    ),
+                    timeout: 0,
+                    icon: 'error',
+                },
+                first,
+            );
+        }
+    }, [health]);
+
+    return health;
+};
+
+export default useHealth;
